Create Lesson3HTML
Create an HTML file with a form so the end user can select a dinner entree: steak, salmon, barbecue pork, or chicken. Modify your PHP receipt program from the Lesson 2 project(be sure to save the file under a new name) so that it uses an if/else statement to automatically select a different beverage to accompany each entree. Use easy to understand variable names and comment your code